Move MAP_LENGTH to a more sensible location
